chore: Updated component folder structure
- Moved Header and Footer components to common folder
- Created Home folder to store its components
- Deleted unused components
