--- src/components/form.tsx
@@ -1,5 +1,0 @@
-import { Widget } from "@typeform/embed-react";
-
-export default function Form() {
-  return <Widget data-tf-opacity="0" id="fiOmcDcX" className="my-form h-full w-full bg-transparent border border-black rounded-2xl overflow-hidden" />
-}
